+++ sequence_layers/jax/convolution2d_test.py
@@ -211,7 +211,7 @@ class Conv2DTest(test_utils.SequenceLayerTest):
         time_padding=time_padding,
         spatial_padding=spatial_padding,
         use_weight_norm=use_weight_norm,
-        dtype=compute_dtype,
+        compute_dtype=compute_dtype,
         param_dtype=param_dtype,
         name='conv2d',
     ).make()
@@ -392,7 +392,7 @@ class Conv2DTransposeTest(test_utils.SequenceLayerTest):
         time_padding=time_padding,
         spatial_padding=spatial_padding,
         use_weight_norm=use_weight_norm,
-        dtype=compute_dtype,
+        compute_dtype=compute_dtype,
         name='transpose_conv2d',
     ).make()
